feat: 테스트 재정리
(cherry picked from commit d9d95aa059ef62fc53bf464d14625a7826b2c3e9)
